pyproject.toml
[tool.poetry]
name = "barfmeister"
version = "2.0"
description = "Barfmeister discord-bot"
authors = ["Mark Boute <bmark0702@gmail.com>"]
license = "n"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7"
python-dotenv = "^0.14.0"
discord = "^1.7.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"